refactoring of Product Plugin
