+++ app/code/Advox/Employees/Api/Data/ProductInterfacePlugin.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Advox\Employees\Api\Data;
+
+use Magento\Catalog\Api\Data\ProductInterface;
+
+interface ProductInterfacePlugin
+{
+    public function afterGetName(ProductInterface $subject, $name): string;
+}
